pyproject.toml
[project]
name = "pyamlconf"
version = "0.0.1"
description = "A simple YAML-based configuration loader for Python"
authors = [
    { name = "Monster Yao", email = "yymmiinngg@163.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pyyaml"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"